+++ CDL/Math/FixedPoint.h
@@ -4,7 +4,7 @@
  *  @author   alcoco
  *  @date     
  *   Created:       15:59:03 08/03/2005
- *   Last Update:   14:58:19 15/06/2005
+ *   Last Update:   00:42:13 01/10/2006
  */
 //========================================================================
 
@@ -18,7 +18,7 @@ namespace CDL
     class DLL_API FixedPoint
     {
         public:
-            static int D_PRECISION;
+            static size_t D_PRECISION;
             typedef long long fpoint_t;
 
         private:
@@ -26,9 +26,9 @@ namespace CDL
             char     m_decBits;
 
         public:
-            FixedPoint(const int &d=D_PRECISION, const fpoint_t &v=0LL);
-            FixedPoint(const int &d, const float &);
-            FixedPoint(const int &d, const double &);
+            FixedPoint(const size_t &d=D_PRECISION, const fpoint_t &v=0LL);
+            FixedPoint(const size_t &d, const float &);
+            FixedPoint(const size_t &d, const double &);
             FixedPoint(const FixedPoint &);
             FixedPoint(const float &);
             FixedPoint(const double &);
